requirements.txt
absl_py>=0.9.0
ml_collections>=0.1.1
nflows>=0.14
numpy>=1.23.1
rdkit>=2022.3.4
torch>=1.12.0+cu113
tqdm>=4.64.0
